This update is a major revamp of the product category/style/characteristics system.
The product table now requires the inclusion of a product_category_id; my feeling is that this is not an onerous requirement, and it will be useful to be able to rely on this information being present.The product_style and product_characteristic_type tables are still associated with product_category. Database triggers have been installed to enforce usage of style and characteristic for a given product category (tests still need to be written though).

The reasons for these changes are that (a) we need to be able to assign a product category in the absence of style or characteristic data, and (b) the previous schema allowed a many-to-many association between product and category (via product_characteristic), which is probably undesirable on a number of levels.